Add Documents Tab to UI
